requirements.txt

python-dotenv>=1.0.1
streamlit>=1.35.0
pydantic>=2.6.0
groq
